@@ -85,7 +85,7 @@ const rule: Rule.RuleModule = {
     // variables should be defined here
     let insideCallback = false;
     let insideElement = false;
-    const source = context.getSourceCode();
+    const source = context.sourceCode;
 
     //----------------------------------------------------------------------
     // Helpers
